fixes to randomutil
