+++ math/src/main/java/com/dumbdogdiner/stickyapi/math/NumberUtil.java
@@ -8,6 +8,8 @@ import com.google.common.base.Preconditions;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.text.CharacterIterator;
+import java.text.StringCharacterIterator;
 import java.util.regex.Pattern;
 
 /**
@@ -196,5 +198,27 @@ public final class NumberUtil {
 					throw new ArithmeticException(); // Somehow Long.compare is broken?? This should be impossible
 			}
 		}
+	}
+
+	/**
+	 * Convert a number of bytes to a human-readable value.
+	 * TODO: How the absolute fuck does this work?
+	 *
+	 * @param bytes the value that should be converted
+	 * @return a human-readable byte value
+	 */
+	public static String bytesToReadable(long bytes) {
+		long absB = bytes == Long.MIN_VALUE ? Long.MAX_VALUE : Math.abs(bytes);
+		if (absB < 1024) {
+			return bytes + " B";
+		}
+		long value = absB;
+		CharacterIterator ci = new StringCharacterIterator("KMGTPE");
+		for (int i = 40; i >= 0 && absB > 0xfffccccccccccccL >> i; i -= 10) {
+			value >>= 10;
+			ci.next();
+		}
+		value *= Long.signum(bytes);
+		return String.format("%.1f %ciB", value / 1024.0, ci.current());
 	}
 }
